Refactor UI components and fix config
