+++ src/app/Auth/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '../(GroupA)/auth/page';
